Selenium/requirements.txt
pytest
selenium>=4.10.0
webdriver-manager
pyyaml
pyautogui
pandas
requests
pytest-html
allure-pytest
python-dotenv